# Conflicts:
#	Backend/User/Infrastructure/UserRepository.cs

+++ Backend/Movie/Domain/Rating.cs
@@ -2,7 +2,7 @@
 
 public class Rating
 {
-    public float AverageRating  { get; set; }
+    public double AverageRating  { get; set; }
     public int Votes { get; set; }
 
     public void NewRating(int newRating)
